backend/requirements.txt
# PipeGenie Backend Dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# MongoDB
motor==3.6.0
beanie==1.26.0

# Redis
redis==5.0.6

# AI / LangChain (keep in sync: google-genai needs langchain-core>=0.2.33)
langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-mistralai==0.1.8
langchain-google-genai==1.0.10

# MilvusDB
pymilvus==2.4.0
setuptools<81
marshmallow<4

# GitHub & HTTP
httpx==0.27.0
PyGithub==2.3.0

# Docker
docker==7.1.0

# Utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Observability / OpenTelemetry (SigNoz)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi

# Testing
pytest==8.3.3
